\name{netPerformance}


\alias{netPerformance}

\title{Net performance of the portfolio}


\description{Displays plot of rebased portfolio performance and summary statistics.}


\usage{   
netPerformance(object, format = "\%Y-\%m-\%d")
}


\arguments{

    \item{object}{
        a list, returned from running the function \code{portfolioSmoothing}.
        }
    \item{format}{
        a character string of the date format used
       }
    

}


\value{A plot of rebased portfolio returns and tables summarising portfolio performance over time}


\keyword{models}

